Add ownsEventLoopGroup flag
Currently, the DiscordClient always takes ownership by tying the event
loop group's lifetime to itself (by shutting it down in the
deinitializer). Usually, however, when an explicit event loop group is
provided, the lifetime is managed externally, therefore we change the
semantics to only shut down the event loop group if created by the
DiscordClient itself.